fix artifacts metrics collection
